+++ wpilibc/athena/include/Joystick.h
@@ -95,7 +95,7 @@ class Joystick : public GenericHID, public ErrorBase {
   bool GetIsXbox() const;
   Joystick::HIDType GetType() const;
   std::string GetName() const;
-  int GetAxisType(uint8_t axis) const;
+  int GetAxisType(uint32_t axis) const;
 
   int GetAxisCount() const;
   int GetButtonCount() const;
